requirements.txt
asgiref==3.6.0
Django==4.1.5
django-crispy-forms==1.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
